Refactor BSOD and color test into utils module
Moved BSOD rendering and rainbow color test functions from main.rs into dedicated utils submodules (bsod.rs and color_test.rs). Updated main.rs to use the new module paths for these features, improving code organization and maintainability.
