+++ kernel/src/utils/mod.rs
@@ -1,0 +1,2 @@
+pub mod bsod;
+pub mod color_test;
